moved simplecopypaste to detection module
